修复一下无法编译的问题

+++ handler/src/test/java/io/netty/handler/stream/ChunkedWriteHandlerTest.java
@@ -223,7 +223,7 @@ public class ChunkedWriteHandlerTest {
         ch.checkException();
         assertTrue(ch.finish());
 
-        assertEquals(0, ch.readOutbound());
+        assertEquals((Integer) 0, ch.readOutbound());
         assertNull(ch.readOutbound());
     }
